backend/requirements/development.txt
-r base.txt

# Development tools
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0
factory-boy>=3.3.0
faker>=21.0.0

# Code quality
black>=23.12.0
isort>=5.13.2
flake8>=6.1.0
pylint>=3.0.3
mypy>=1.7.1

# Security
django-stubs>=4.2.7
